refactor(napi/oxlint): move external linter construction into own module (#13710)
Pure refactor. Move `wrap_*` functions into separate `external_linter.rs` module, and add a `create_external_linter` function which wraps both JS callbacks.

This is necessary so we can feature gate loading the module (next PR #13711).
